Add hydra-gen-jwks
Hydra has an apparent bug where if the /.well-known/jwks.json endpoint is
queried multiple times simultaneous or in quick succession and the JWKS doesn't
exist yet, it will try to generate the JWKS twice, changing which keys are used
to sign valid JWTs. This makes the valid JWTs return as unauthorized when they
in fact are authorized.

To mitigate this, the hydra-gen-jwks service generates the JWKS once when Hydra
starts up. The services that need the JWKS (i.e. SMD and BSS) depend on this
service completing successfully. That way, the JWKS is already generated when
they ask for it and the race condition is avoided.