Add cloud-init service
This commit adds a new service called cloud-init to the quickstart deployment. The cloud-init service is responsible for delivering cloud-init payloads based on the contents of SMD. It is dependent on the smd service and uses the CLOUD_INIT_ACCESS_TOKEN environment variable for authentication.

There are correspoonding changes to haproxy where I have also made improvements to logging and possible race conditions for starting services.

The README has been updated as well to cover the addition of this container