quickstart(bss,tpm-manager): update to new BSS notifier system
When a node requests its bootscript, BSS can now initiate a POST request
to and endpoint of the user's choice — in this case, to the TPM manager
daemon. This notification includes the booting node's IP address, which
is passed to Ansible as inventory to be configured with a TPM key.